+++ include/ParallaxGen/ParallaxGen.hpp
@@ -39,7 +39,7 @@ public:
 
 private:
     // processes a NIF file (enable parallax if needed)
-	void processNIF(std::filesystem::path nif_file, std::vector<std::filesystem::path>& heightMaps, std::vector<std::filesystem::path>& complexMaterialMaps);
+	void processNIF(const std::filesystem::path& nif_file, std::vector<std::filesystem::path>& heightMaps, std::vector<std::filesystem::path>& complexMaterialMaps);
 	// enables complex material on a shape in a NIF
 	bool enableComplexMaterialOnShape(nifly::NifFile& nif, nifly::NiShape* shape, nifly::NiShader* shader, const std::string& search_prefix);
 	// enables parallax on a shape in a NIF