+++ include/ParallaxGenUtil/ParallaxGenUtil.hpp
@@ -11,13 +11,13 @@
 namespace ParallaxGenUtil
 {
 	// reads ini file into ptree
-	boost::property_tree::ptree readINIFile(const std::filesystem::path& ini_path, const bool required);
+	boost::property_tree::wptree readINIFile(const std::filesystem::path& ini_path, const bool required = false, const bool logging = false);
 
 	// opens a file handle for a file
 	std::ifstream openFileHandle(const std::filesystem::path& file_path, const bool required);
 
 	// merges two ptrees
-	void mergePropertyTrees(boost::property_tree::ptree& pt1, const boost::property_tree::ptree& pt2);
+	void mergePropertyTrees(boost::property_tree::wptree& pt1, const boost::property_tree::wptree& pt2);
 
 	// gets the system path for a folder (from windows.h)
 	std::filesystem::path getSystemPath(const GUID& folder_id);
@@ -27,6 +27,13 @@ namespace ParallaxGenUtil
 
 	// converts fs::path to a lowercase variant (for case-insensitive comparison)
 	void pathLower(std::filesystem::path& path);
+
+	// converts a string to a wstring
+	std::wstring convertToWstring(const std::string str);
+
+	std::string wstring_to_utf8(const std::wstring& str);
+
+	std::vector<std::byte> getFileBytes(const std::filesystem::path& file_path);
 
 	// concatenates two vectors without duplicates
 	template <typename T>