+++ src/BethesdaDirectory/BethesdaDirectory.cpp
@@ -7,6 +7,8 @@
 #include <fstream>
 #include <boost/algorithm/string.hpp>
 #include <binary_io/binary_io.hpp>
+#include <boost/iostreams/device/mapped_file.hpp>
+#include <boost/filesystem.hpp>
 
 // BSA Includes
 #include <cstdio>
@@ -23,7 +25,7 @@ BethesdaDirectory::BethesdaDirectory(BethesdaGame& bg, const bool logging)
 
 	if(this->logging) {
 		// Log starting message
-		spdlog::info("Opening Data Folder \"{}\"", bg.getGameDataPath().string());
+		spdlog::info(L"Opening Data Folder \"{}\"", bg.getGameDataPath().wstring());
 	}
 
 	// Assign instance vars
@@ -50,28 +52,31 @@ map<fs::path, shared_ptr<BethesdaDirectory::BSAFile>> BethesdaDirectory::getFile
 
 vector<std::byte> BethesdaDirectory::getFile(const fs::path rel_path) const
 {
+	// check if key exists in map
+	if (fileMap.find(rel_path) == fileMap.end()) {
+		if (logging) {
+			spdlog::error(L"File not found in file map: {}", rel_path.wstring());
+			return vector<std::byte>();
+		}
+		else {
+			throw runtime_error("File not found in file map");
+		}
+	}
+
 	// find bsa/loose file to open
 	shared_ptr<BSAFile> bsa_struct = fileMap.at(rel_path);
 	if (bsa_struct == nullptr) {
 		if (logging) {
-			spdlog::trace("Reading loose file from BethesdaDirectory: {}", rel_path.string());
+			spdlog::trace(L"Reading loose file from BethesdaDirectory: {}", rel_path.wstring());
 		}
 
-		// this is a loose file
 		fs::path file_path = data_dir / rel_path;
-		binary_io::file_istream fis(file_path);
 
-		// create buffer to store bytes
-		//todo: is there a way to avoid the span conversion here?
-		vector<std::byte> buffer(fs::file_size(file_path));
-		span<std::byte> span(buffer);
-		fis.read_bytes(span);
-
-		return buffer;
+		return getFileBytes(file_path);
 	}
 	else {
 		if (logging) {
-			spdlog::trace("Reading BSA file from BethesdaDirectory: {}", rel_path.string());
+			spdlog::trace(L"Reading BSA file from BethesdaDirectory: {}", rel_path.wstring());
 		}
 
 		// this is a bsa archive file
@@ -87,27 +92,43 @@ vector<std::byte> BethesdaDirectory::getFile(const fs::path rel_path) const
 			return s.rdbuf();
 		}
 		else {
-			throw runtime_error("File not found in BSA archive. This can only happen if the file was deleted after this object was constructed.");
+			if (logging) {
+				spdlog::error(L"File not found in BSA archive: {}", rel_path.wstring());
+			}
+			else {
+				throw runtime_error("File not found in BSA archive");
+			}
 		}
 	}
+
+	return vector<std::byte>();
 }
 
-vector<fs::path> BethesdaDirectory::findFilesBySuffix(const string_view suffix, const vector<string>& path_blocklist) const
+vector<fs::path> BethesdaDirectory::findFilesBySuffix(const string_view suffix, const vector<wstring>& path_blocklist) const
 {
 	// find all keys in fileMap that match pattern
 	vector<fs::path> found_files;
 
 	// loop through filemap and match keys
 	for (const auto& [key, value] : fileMap) {
+
+		if (logging) {
+			spdlog::trace(L"Checking file for suffix {} match: {}", convertToWstring(string(suffix)), key.wstring());
+		}
+
         if (boost::algorithm::ends_with(key.wstring(), suffix)) {
 			// check if any component of the path is in the blocklist
 			bool block = false;
-			for (string block_path : path_blocklist) {
+			for (wstring block_path : path_blocklist) {
 				for (const auto& component : key) {
-					if (block_path == component.string()) {
+					if (block_path == component.wstring()) {
 						block = true;
 						break;
 					}
+				}
+				
+				if (block) {
+					break;
 				}
 			}
 
@@ -116,7 +137,7 @@ vector<fs::path> BethesdaDirectory::findFilesBySuffix(const string_view suffix, 
 			}
 
 			if (logging) {
-				spdlog::trace("Matched file by suffix: {}", key.string());
+				spdlog::trace(L"Matched file by suffix: {}", key.wstring());
 			}
 
 			found_files.push_back(key);
@@ -133,10 +154,10 @@ void BethesdaDirectory::addBSAFilesToMap()
 	}
 
 	// Get list of BSA files
-	vector<string> bsa_files = getBSAPriorityList();
+	vector<wstring> bsa_files = getBSAPriorityList();
 
 	// Loop through each BSA file
-	for (string bsa_name : bsa_files) {
+	for (wstring bsa_name : bsa_files) {
 		// add bsa to file map
 		addBSAToFileMap(bsa_name);
 	}
@@ -159,16 +180,20 @@ void BethesdaDirectory::addLooseFilesToMap()
 				continue;
 			}
 
+			if (logging) {
+				spdlog::trace(L"Adding loose file to map: {}", relative_path.wstring());
+			}
+
 			fileMap[relative_path] = nullptr;
 		}
 	}
 }
 
-void BethesdaDirectory::addBSAToFileMap(const string& bsa_name)
+void BethesdaDirectory::addBSAToFileMap(const wstring& bsa_name)
 {
 	if(logging) {
 		// log message
-		spdlog::debug("Adding files from {} to file map.", bsa_name);
+		spdlog::debug(L"Adding files from {} to file map.", bsa_name);
 	}
 
 	bsa::tes4::archive bsa_obj;
@@ -177,7 +202,7 @@ void BethesdaDirectory::addBSAToFileMap(const string& bsa_name)
 	// skip BSA if it doesn't exist (can happen if it's in the ini but not in the data folder)
 	if (!fs::exists(bsa_path)) {
 		if(logging) {
-			spdlog::warn("Skipping BSA {} because it doesn't exist", bsa_path.string());
+			spdlog::warn(L"Skipping BSA {} because it doesn't exist", bsa_path.wstring());
 		}
 		return;
 	}
@@ -212,38 +237,42 @@ void BethesdaDirectory::addBSAToFileMap(const string& bsa_name)
 				continue;
 			}
 
+			if (logging) {
+				spdlog::trace(L"Adding file from BSA {} to file map: {}", bsa_name, cur_path.wstring());
+			}
+
 			// add to filemap
 			fileMap[cur_path] = bsa_shared_ptr;
 		}
 	}
 }
 
-vector<string> BethesdaDirectory::getBSAPriorityList() const
+vector<wstring> BethesdaDirectory::getBSAPriorityList() const
 {
 	// get bsa files not loaded from esp (also initializes output vector)
-	vector<string> out_bsa_order = getBSAFilesFromINIs();
+	vector<wstring> out_bsa_order = getBSAFilesFromINIs();
 
 	// get esp priority list
-	const vector<string> load_order = getPluginLoadOrder(true);
+	const vector<wstring> load_order = getPluginLoadOrder(true);
 
 	// list BSA files in data directory
-	const vector<string> all_bsa_files = getBSAFilesInDirectory();
+	const vector<wstring> all_bsa_files = getBSAFilesInDirectory();
 
 	//loop through each esp in the priority list
-	for (string plugin : load_order) {
+	for (wstring plugin : load_order) {
 		// add any BSAs to list
-		vector<string> cur_found_bsas = findBSAFilesFromPluginName(all_bsa_files, plugin);
+		vector<wstring> cur_found_bsas = findBSAFilesFromPluginName(all_bsa_files, plugin);
 		concatenateVectorsWithoutDuplicates(out_bsa_order, cur_found_bsas);
 	}
 
 	if(logging) {
 		// log output
-		string bsa_list_str = boost::algorithm::join(out_bsa_order, ",");
-		spdlog::trace("BSA Load Order: {}", bsa_list_str);
+		wstring bsa_list_str = boost::algorithm::join(out_bsa_order, ",");
+		spdlog::trace(L"BSA Load Order: {}", bsa_list_str);
 
-		for (string bsa : all_bsa_files) {
+		for (wstring bsa : all_bsa_files) {
 			if (find(out_bsa_order.begin(), out_bsa_order.end(), bsa) == out_bsa_order.end()) {
-				spdlog::warn("BSA file {} not loaded by any plugin or INI.", bsa);
+				spdlog::warn(L"BSA file {} not loaded by any plugin or INI.", bsa);
 			}
 		}
 	}
@@ -251,10 +280,10 @@ vector<string> BethesdaDirectory::getBSAPriorityList() const
 	return out_bsa_order;
 }
 
-vector<string> BethesdaDirectory::getPluginLoadOrder(const bool trim_extension) const
+vector<wstring> BethesdaDirectory::getPluginLoadOrder(const bool trim_extension) const
 {
 	// initialize output vector. Stores
-	vector<string> output_lo;
+	vector<wstring> output_lo;
 
 	// set path to loadorder.txt
 	fs::path lo_file = getGameAppdataPath() / "loadorder.txt";
@@ -265,15 +294,24 @@ vector<string> BethesdaDirectory::getPluginLoadOrder(const bool trim_extension) 
 			exitWithUserInput(1);
 		}
 		else {
-			throw runtime_error("loadorder.txt not found.");
+			throw runtime_error("loadorder.txt not found");
 		}
 	}
 
 	// open file
-	ifstream f = openFileHandle(lo_file, true);
+	wifstream f(lo_file, true);
+	if (!f.is_open()) {
+		if (logging) {
+			spdlog::critical("Unable to open loadorder.txt");
+			exitWithUserInput(1);
+		}
+		else {
+			throw runtime_error("Unable to open loadorder.txt");
+		}
+	}
 
 	// loop through each line of loadorder.txt
-	string line;
+	wstring line;
 	while (getline(f, line)) {
 		// Ignore lines that start with '#', which are comment lines
 		if (line.empty() || line[0] == '#') {
@@ -293,28 +331,29 @@ vector<string> BethesdaDirectory::getPluginLoadOrder(const bool trim_extension) 
 	f.close();
 
 	if(logging) {
-		string load_order_str = boost::algorithm::join(output_lo, ",");
-		spdlog::debug("Plugin Load Order: {}", load_order_str);
+		wstring load_order_str = boost::algorithm::join(output_lo, ",");
+		spdlog::debug(L"Plugin Load Order: {}", load_order_str);
 	}
 
 	// return output
 	return output_lo;
 }
 
-vector<string> BethesdaDirectory::getBSAFilesFromINIs() const
+vector<wstring> BethesdaDirectory::getBSAFilesFromINIs() const
 {
 	// output vector
-	vector<string> bsa_files;
+	vector<wstring> bsa_files;
 
 	// get ini properties
-	boost::property_tree::ptree pt_ini = getINIProperties();
+	boost::property_tree::wptree pt_ini = getINIProperties();
 
 	// loop through each archive field in the INI files
 	for (string field : ini_bsa_fields) {
-		string cur_val;
+		wstring cur_val;
 
 		try {
-			cur_val = pt_ini.get<string>("Archive." + field);
+			wstring field_name = convertToWstring("Archive." + field);
+			cur_val = pt_ini.get<wstring>(field_name);
 		}
 		catch (const exception& e) {
 			spdlog::info("Unable to find {} in [Archive] section in game ini: {}: Ignoring.", field, e.what());
@@ -322,10 +361,10 @@ vector<string> BethesdaDirectory::getBSAFilesFromINIs() const
 		}
 
 		// split into components
-		vector<string> ini_components;
+		vector<wstring> ini_components;
 		boost::split(ini_components, cur_val, boost::is_any_of(","));
 
-		for (string bsa : ini_components) {
+		for (wstring bsa : ini_components) {
 			// remove leading/trailing whitespace
 			boost::trim(bsa);
 
@@ -338,9 +377,9 @@ vector<string> BethesdaDirectory::getBSAFilesFromINIs() const
 	return bsa_files;
 }
 
-vector<string> BethesdaDirectory::getBSAFilesInDirectory() const
+vector<wstring> BethesdaDirectory::getBSAFilesInDirectory() const
 {
-	vector<string> bsa_files;
+	vector<wstring> bsa_files;
 
 	for (const auto& entry : fs::directory_iterator(this->data_dir)) {
 		if (entry.is_regular_file()) {
@@ -351,35 +390,35 @@ vector<string> BethesdaDirectory::getBSAFilesInDirectory() const
 			}
 
 			// add to output
-			bsa_files.push_back(entry.path().filename().string());
+			bsa_files.push_back(entry.path().filename().wstring());
 		}
 	}
 
 	return bsa_files;
 }
 
-vector<string> BethesdaDirectory::findBSAFilesFromPluginName(const vector<string>& bsa_file_list, const string& plugin_prefix) const
+vector<wstring> BethesdaDirectory::findBSAFilesFromPluginName(const vector<wstring>& bsa_file_list, const wstring& plugin_prefix) const
 {
-	vector<string> bsa_files_found;
+	vector<wstring> bsa_files_found;
 
-	for (string bsa : bsa_file_list) {
+	for (wstring bsa : bsa_file_list) {
 		if (bsa.starts_with(plugin_prefix)) {
-			if (bsa == plugin_prefix + ".bsa") {
+			if (bsa == plugin_prefix + L".bsa") {
 				// load bsa with the plugin name before any others
 				bsa_files_found.insert(bsa_files_found.begin(), bsa);
 				continue;
 			}
 
 			// skip any BSAs that may start with the prefix but belong to a different plugin
-			string after_prefix = bsa.substr(plugin_prefix.length());
+			wstring after_prefix = bsa.substr(plugin_prefix.length());
 
 			// todo: Is this actually how the game handles BSA files? Example: 3DNPC0.bsa, 3DNPC1.bsa, 3DNPC2.bsa are loaded,
 			// todo: but 3DNPC - Textures.bsa is also loaded, whats the logic there?
-			if (after_prefix.starts_with(" ") && !after_prefix.starts_with(" -")) {
+			if (after_prefix.starts_with(L" ") && !after_prefix.starts_with(L" -")) {
 				continue;
 			}
 
-			if (!after_prefix.starts_with(" ") && !isdigit(after_prefix[0])) {
+			if (!after_prefix.starts_with(L" ") && !isdigit(after_prefix[0])) {
 				continue;
 			}
 
@@ -402,7 +441,7 @@ bool BethesdaDirectory::file_allowed(const fs::path file_path) const
 	return true;
 }
 
-boost::property_tree::ptree BethesdaDirectory::getINIProperties() const
+boost::property_tree::wptree BethesdaDirectory::getINIProperties() const
 {
 	// get document path
 	fs::path doc_path = getGameDocumentPath();
@@ -411,8 +450,8 @@ boost::property_tree::ptree BethesdaDirectory::getINIProperties() const
 	fs::path ini_path = doc_path / BethesdaGame::INILocations.at(game_type).ini;
 	fs::path custom_ini_path = doc_path / BethesdaGame::INILocations.at(game_type).ini_custom;
 
-	boost::property_tree::ptree pt_ini = readINIFile(ini_path, true);
-	boost::property_tree::ptree pt_custom_ini = readINIFile(custom_ini_path, false);
+	boost::property_tree::wptree pt_ini = readINIFile(ini_path, false);
+	boost::property_tree::wptree pt_custom_ini = readINIFile(custom_ini_path, false);
 
 	mergePropertyTrees(pt_ini, pt_custom_ini);